requirement.txt
asgiref==3.5.1
backports.zoneinfo==0.2.1
Django==4.0.4
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
PyJWT==2.3.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
